Keep the true frame rate instead of truncating it to an integer
Seven call sites read int(vidcap.get(cv2.CAP_PROP_FPS)): _directograms, _flow
twice, _history, _impacts, _warp and _videoadjust. On NTSC-rate footage that is
int(29.97) == 29, so every time and frequency they derived was 3.2 per cent low,
and _flow and _history wrote the truncated value into the output file's declared
rate, turning a 29.97 fps input into a 29 fps file. The docs carried it as a
known defect awaiting a fix.

Every use was checked to be safe with a non-integer rate before changing it:
divisions, cv2.VideoWriter, ffmpeg -r and librosa sr all take a float, and
_impacts already wrapped int() where it needed a whole number.

Pinned by a guard test that fails if any module reads the rate through int()
again, and by one that shows 29.97 surviving where int() gave 29. The docs note
is rewritten from a standing warning into a record of the fix, saying plainly
that figures produced before it are not comparable with figures after it on
non-integer-rate footage.

592 tests pass.

+++ musicalgestures/_directograms.py
@@ -90,7 +90,7 @@ def mg_directograms(self, title: str | None = None, filtertype: str = 'Adaptativ
     _ensure_numba()  # JIT-compile the directogram kernels on first use
 
     vidcap = cv2.VideoCapture(filename)
-    fps = int(vidcap.get(cv2.CAP_PROP_FPS))
+    fps = float(vidcap.get(cv2.CAP_PROP_FPS))
     length = int(vidcap.get(cv2.CAP_PROP_FRAME_COUNT))
 
     pb = MgProgressbar(total=length, prefix='Rendering directogram:')
